RFC: Move DARKTIME calc to dodarktime.c
RFC: Move some BLEVCORR stuff out.
