+++ pkg/acs/include/acs.h
@@ -2,10 +2,10 @@
 #define ACS_INCL
 
 /* acs.h generic header for calacs */
-/*      */
-# include <stdio.h>             /* To insure that FILE is defined for TrlPtr */
-# include "imphttab.h"
-# include "hstcal.h"
+
+#include <stdio.h>             /* To insure that FILE is defined for TrlPtr */
+#include "imphttab.h"
+#include "hstcal.h"
 #include "trlbuf.h"
 
 # define ACS_CBUF           24  /* small buffer for e.g. rootname */
